Move FutureWeatherDao to dao/
