+++ app/src/main/java/com/example/bcsd_weather/di/KoinModule.kt
@@ -4,7 +4,7 @@ package com.example.bcsd_weather.di
 import android.app.Application
 import androidx.room.Room
 import com.example.bcsd_weather.FutureWeatherViewModel
-import com.example.bcsd_weather.data.datasource.local.FutureWeatherDao
+import com.example.bcsd_weather.data.dao.FutureWeatherDao
 import com.example.bcsd_weather.data.db.FutureWeatherDatabase
 import com.example.bcsd_weather.data.repository.FutureWeatherRepositoryImpl
 import com.example.bcsd_weather.domain.repository.FutureWeatherRepository
